==> 01 Data Prep.Rmd <==
---
title: "Data Prep"
author: "Xiang XU"
date: "November 20, 2018"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
pacman::p_load(jsonlite)
setwd("E:/MSSP/MA678/Midproj_Yelp")
```


##read business data
```{r}
#clear memory
rm(list = ls())

#create business dataframe
business_rawdata <- lapply(readLines("dataset/yelp_academic_dataset_business.json"),jsonlite::fromJSON)
print("original data loaded")


#business_id
business_id <- sapply(business_rawdata, function(x){x$business_id})

#extract business hour info
day_f <-  c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday","Saturday", "Sunday")

fill_df_business_hours  <-  function(row_business){
  row = rep(NA, times = 7)
  for (i in 1:7){
    if (!is.null(row_business$hours[[day_f[i]]])){
      row[i] = row_business$hours[[day_f[i]]]
    }
  }
  return(row)
}

# create a temporary matrix to store the business hours

hours_temp  <-  sapply(business_rawdata, fill_df_business_hours)
hours_temp  <-  t(hours_temp)

# function to split opening and closing time of each day

split_open_close <-  function(daily_hours_temp){
  if (is.na(daily_hours_temp)){
    daily_open_close = c(NA, NA)
  } else{
    daily_open_close = strsplit(daily_hours_temp, split = "-")
  }
return(daily_open_close)
}



# create a matrix with opening and closing time of each day

vector_temp = unlist(sapply(hours_temp[, 1], split_open_close))
open_close = matrix(vector_temp, ncol = 2, byrow = T)
for (i in 2:7){
  vector_temp = unlist(sapply(hours_temp[, i], split_open_close))
  open_close = cbind(open_close, matrix(vector_temp, ncol = 2, byrow = T))
}

# create col names for the dataframe regarding business hours
# M = Monday, T = Tuesday, W = Wednesday, TH = Thursday
# F = Friday, S = Saturday, SU = Sunday
# O = Opening, C = Closing

day = c("M_", "T_", "W_", "TH_", "F_", "S_", "SU_")
cname_business_hours = c("business_id", as.vector(sapply(day, paste0, c("O", "C"))))


# create the dataframe regarding business hours
df_business_hours = data.frame(business_id, open_close, stringsAsFactors = F)
colnames(df_business_hours) = cname_business_hours
print("dataframe for business hours created")


# save the dataframe regarding business hours in an Rdata file

save(df_business_hours, file = "dataset/business_hours.Rdata")

print("dataframe for business hours saved to an Rdata file")



# remove temporary variables

rm(list = c("hours_temp", "open_close", "cname_business_hours",

            "day", "day_f", "i", "vector_temp"))





############################################

# create a dataframe related to business attributes

# create a list of all possible attributes

all_attributes = names(unlist(business_rawdata[[1]]$attributes))

for (i in 2:length(business_rawdata)){

  attributes_temp = names(unlist(business_rawdata[[i]]$attributes))

  all_attributes = union(all_attributes, attributes_temp)

}

rm("attributes_temp")



# a function to find nested attributes

is.nested = function(att){

  return(grepl(att, pattern = "\\."))

}



# a function to retrieve information

retrieve_att = function(att){

  if (is.nested(att)){

    name1 = strsplit(att, "\\.")[[1]][1]

    name2 = strsplit(att, "\\.")[[1]][2]

    ra1 = function(row_business){

      if(is.null(row_business$attributes[[name1]][[name2]])) {

        NA

      } else row_business$attributes[[name1]][[name2]]

    }

    return(sapply(business_rawdata, ra1))

  } else{

    ra2 = function(row_business){

      if(is.null(row_business$attributes[[att]])) {

        NA

      } else row_business$attributes[[att]]

    }

    return(sapply(business_rawdata, ra2))

  }

}



# create the dataframe related to business attributes

df_business_attributes = data.frame(retrieve_att(all_attributes[1]))

for (i in 2:length(all_attributes)){

  attribute = retrieve_att(all_attributes[i])

  df_business_attributes = data.frame(df_business_attributes, attribute)

  print(i)

}

df_business_attributes = data.frame(business_id, df_business_attributes,

                                    stringsAsFactors = F)

colnames(df_business_attributes) = c("business_id", all_attributes)

rm(list = c("attribute", "i", "all_attributes"))

print("dataframe for business attributes created")



# save the dataframe to an Rdata file

save(df_business_attributes, file = "dataset/business_attributes.Rdata")

print("dataframe for business attributes saved to an Rdata file")





###########################################

# create a dataframe related to basic business information

# create a list of all columns containing basic information

b_info = names(business_rawdata[[1]])

for (i in 1:length(business_rawdata)){

  b_info_temp = names(business_rawdata[[i]])

  b_info = union(b_info, b_info_temp)

}

rm(list = c("b_info_temp", "i"))

b_info = b_info[-grep(b_info, pattern = "hours|attributes" )]

b_info = b_info[c(-1, -13)]



# converting category information into a string

collect_category = function(row_business){

  if (length(row_business$categories) > 0){

    return(paste(row_business$categories, collapse = " * "))

  } else {

    return(NA)

  }

}

categories = sapply(business_rawdata, collect_category)



# a function to retrieve basic information

retrieve_b_info = function(binfo){

  rb = function(row_business){

    if(is.null(row_business[[binfo]])) {

      NA

    } else row_business[[binfo]]

  }

  return(sapply(business_rawdata, rb))

}



# create the dataframe related to basic business information

df_business_info = data.frame(retrieve_b_info(b_info[1]), stringsAsFactors = F)

for (i in 2:length(b_info)){

  information = retrieve_b_info(b_info[i])

  df_business_info = data.frame(df_business_info, information, 

                                stringsAsFactors = F)

}

colnames(df_business_info) = b_info

df_business_info = data.frame(business_id, df_business_info, categories,

                              stringsAsFactors = F)

df_business_info$state = as.factor(df_business_info$state)

df_business_info$is_open = as.logical(df_business_info$is_open)

rm(list = c("b_info", "i", "information", "categories"))

print("dataframe for basic business information created")



# save the dataframe for basic information to an Rdata file

save(df_business_info, file = "dataset/business_info.Rdata")

print("dataframe for basic information saved to Rdata file")
```



##read check in data


```{r}
# clear memory

rm(list = ls())

# load packages

library(jsonlite)



# load the original data

checkin_data_original = lapply(readLines("dataset/yelp_academic_dataset_checkin.json"), jsonlite::fromJSON)

print("original data loaded")



###########################################

# function used to fill the dataframe

day_f = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday")

hour_f = rep(NA, times = 24)



for (h in 1:24){

  hour_f[h] = paste0(h-1, ":00")

}



fill_df_checkin = function(row_checkin){

  row = rep(NA, times = length(cnames_checkin))

  row[1] = row_checkin$business_id

  for (i in 1:7){

    for (j in 1:24){

      if (!is.null(row_checkin$time[[day_f[i]]][[hour_f[j]]])){

        row[(i-1)*24 + j + 1] = row_checkin$time[[day_f[i]]][[hour_f[j]]]

      }

    }

  }

  return(row)

}



# create col names for the dataframe

# M = Monday, T = Tuesday, W = Wednesday, TH = Thursday

# F = Friday, S = Saturday, SU = Sunday

day = c("M_", "T_", "W_", "TH_", "F_", "S_", "SU_")

dayhour = as.vector(sapply(day, paste0, 0:23))

cnames_checkin = c("business_id", dayhour)



# create the dataframe

df_checkin = sapply(checkin_data_original, fill_df_checkin)

df_checkin = t(df_checkin)

df_checkin = data.frame(df_checkin, stringsAsFactors = F)

for (i in 2:ncol(df_checkin)){

  df_checkin[, i] = as.numeric(df_checkin[, i])

}

colnames(df_checkin) = cnames_checkin

print("check in dataframe prepared")



# save the dataframe to an rdata file

save(df_checkin, file = "dataset/checkin.Rdata")

print("file saved")
```



##read photo data

```{r}
# clear memory

rm(list = ls())


# load the original data

photos_data_original = lapply(readLines("dataset/yelp_academic_dataset_photo.json"), fromJSON)

print("original data loaded")





##########################################

# create the dataframe

df_photos = sapply(photos_data_original, unlist)

df_photos = t(df_photos)

df_photos = data.frame(df_photos, stringsAsFactors = F)

df_photos$label = as.factor(df_photos$label)

print("dataframe created")



# save photos dataframe to an rdata file

save(df_photos, file = "dataset/photos.Rdata")

print("Rdata file saved")
```

#read review data

```{r}
# clear memory

rm(list = ls())



# load packages

library(jsonlite)

library(lubridate)



# load the original data

txt = readLines("dataset/yelp_academic_dataset_review.json")

print("original review data loaded")



# number of chunks for initial processing

n_chunk = 1000



# number of lines to read per chunk except the last one

n_lines = round(length(txt)/n_chunk)



# current chunk of data

c_chunk = txt[1:n_lines]





################################################

# process the first chunk of data

data_chunk = lapply(c_chunk, fromJSON)



# create col names for the dataframe

cnames = names(data_chunk[[1]])



# function to retrieve review information

retrieve_r_info = function(cname){

  rr = function(row_review){

    if(is.null(row_review[[cname]])){

      NA

    } else row_review[[cname]]

  }

  return(sapply(data_chunk, rr))

}



# create the data frame for review data

df_review = data.frame(retrieve_r_info(cnames[1]), stringsAsFactors = F)

for (i in 2:length(cnames)){

  information = retrieve_r_info(cnames[i])

  df_review = data.frame(df_review, information, stringsAsFactors = F)

}

colnames(df_review) = cnames

rm("information")

print("chunk 1 processing finished")





################################################

## process the rest of the data

for (i in 2:n_chunk){

  

  # load data

  if (i < n_chunk){

    c_chunk = txt[seq(from = (i-1)*n_lines+1, to = i*n_lines)]

  } else{

    c_chunk = txt[seq(from = (i-1)*n_lines+1, to = length(txt))]

  }

  data_chunk = lapply(c_chunk, fromJSON)

  

  # data frame for review data

  df_review_temp = data.frame(retrieve_r_info(cnames[1]), stringsAsFactors = F)

  for (j in 2:length(cnames)){

    information = retrieve_r_info(cnames[j])

    df_review_temp = data.frame(df_review_temp, information,

                                stringsAsFactors = F)

  }

  colnames(df_review_temp) = cnames

  df_review = rbind(df_review, df_review_temp)

  rm(list = c("df_review_temp", "information"))

  print(paste("chunk", i, "processing finished"))

}

df_review$date = ymd(df_review$date)



rm(list = c("c_chunk", "data_chunk", "i", "j"))

print("dataframe for review data created")

save(df_review, file = "dataset/review.Rdata")

print("dataframe for review data saved to an Rdata file")
```



##read tip data

```{r}
# clear memory

rm(list = ls())



# load packages

library(jsonlite)

library(lubridate)



## SET YOUR WORKING DIRECTORY HERE WITH setwd() ##



# load the original data

tip_data_original = lapply(readLines("dataset/yelp_academic_dataset_tip.json"), fromJSON)

print("original data loaded")





######################################

# create the dataframe

df_tip = sapply(tip_data_original, unlist)

df_tip = t(df_tip)

df_tip = data.frame(df_tip, stringsAsFactors = F)

df_tip$date = ymd(df_tip$date)

df_tip$likes = as.numeric(df_tip$likes)

print("dataframe created")



# save tip dataframe to an rdata file

save(df_tip, file = "dataset/tip.Rdata")

print("Rdata file saved")
```


##read user data
```{r}
## This file will create and save 3 dataframes

## dataframe 1: basic information including # of elite years & # of friends

## dataframe 2: detailed information about friends

## dataframe 3: detailed information about elite years



# clear memory

rm(list = ls())



# load packages

library(jsonlite)

library(lubridate)



# load the original data

txt = readLines("dataset/yelp_academic_dataset_user.json")

print("original user data loaded")



# number of chunks for initial processing

n_chunk = 1000



# number of lines to read per chunk except the last one

n_lines = round(length(txt)/n_chunk)



# current chunk of data

c_chunk = txt[1:n_lines]





################################################

# process the first chunk of data

data_chunk = lapply(c_chunk, jsonlite::fromJSON)



## functions to extract information about friends

# find number of friends

find_n_friends = function(row_user){

  return(length(row_user$friends)) 

}



# find id of friends

find_id_friends = function(row_user){

  return(paste(row_user$friends, collapse = " * "))

}



n_friends = sapply(data_chunk, find_n_friends)

friends_id = sapply(data_chunk, find_id_friends)



## function sto extract information about elite status

# find number of years of elite status

find_n_elite = function(row_user){

  return(length(row_user$elite))

}



# find the earlist year of elite status

find_min_elite = function(row_user){

  if (length(row_user$elite) > 0){

    return(min(row_user$elite))

  } else{

    return(NA)

  }

}



# find the latest year of elite status

find_max_elite = function(row_user){

  if (length(row_user$elite) > 0){

    return(max(row_user$elite))

  } else{

    return(NA)

  }

}



# find all elite years

find_elite_years = function(row_user){

  if (length(row_user$elite) > 0){

    return(paste(row_user$elite, collapse = ", "))

  } else{

    return(NA)

  }

} 



n_elite = sapply(data_chunk, find_n_elite)

min_elite = sapply(data_chunk, find_min_elite)

max_elite = sapply(data_chunk, find_max_elite)

elite_years = sapply(data_chunk, find_elite_years)



# a function to retrieve basic user information

cnames = names(data_chunk[[1]])[-c(5, 10)]

retrieve_u_info = function(cname){

  ru = function(row_user){

    if(is.null(row_user[[cname]])) {

      NA

    } else row_user[[cname]]

  }

  return(sapply(data_chunk, ru))

}



# create the dataframe related to basic business information

df_user_info = data.frame(retrieve_u_info(cnames[1]), stringsAsFactors = F)

for (i in 2:length(cnames)){

  information = retrieve_u_info(cnames[i])

  df_user_info = data.frame(df_user_info, information,

                            stringsAsFactors = F)

}

colnames(df_user_info) = cnames

df_user_info = data.frame(df_user_info, n_friends, n_elite,

                          stringsAsFactors = F)

rm("information")

print("Chunk 1 processing finished")





################################################

## process the rest of the data

for (i in 2:n_chunk){

  

  # load data

  if (i < n_chunk){

    c_chunk = txt[seq(from = (i-1)*n_lines+1, to = i*n_lines)]

  } else{

    c_chunk = txt[seq(from = (i-1)*n_lines+1, to = length(txt))]

  }

  data_chunk = lapply(c_chunk, fromJSON)

  

  # information about friends

  n_friends = sapply(data_chunk, find_n_friends)

  friends_id_temp = sapply(data_chunk, find_id_friends)

  friends_id = c(friends_id, friends_id_temp)

 # rm("friends_id_temp")

  

  # information about elite status

  n_elite = sapply(data_chunk, find_n_elite)

  min_elite_temp = sapply(data_chunk, find_min_elite)

  min_elite = c(min_elite, min_elite_temp)

  max_elite_temp = sapply(data_chunk, find_max_elite)

  max_elite = c(max_elite, max_elite_temp)

  elite_years_temp = sapply(data_chunk, find_elite_years)

  elite_years = c(elite_years, elite_years_temp)

 # rm(list = c("min_elite_temp", "max_elite_temp", "elite_years_temp"))

  

  # dataframe related to basic business information

  df_user_info_temp = data.frame(retrieve_u_info(cnames[1]), stringsAsFactors = F)

  for (j in 2:length(cnames)){

    information = retrieve_u_info(cnames[j])

    df_user_info_temp = data.frame(df_user_info_temp, information,

                                   stringsAsFactors = F)

  }

  colnames(df_user_info_temp) = cnames

  df_user_info_temp = data.frame(df_user_info_temp, n_friends, n_elite,

                                 stringsAsFactors = F)

  df_user_info = rbind(df_user_info, df_user_info_temp)

  #rm(list = c("df_user_info_temp", "information"))

  print(paste("chunk", i, "processing finished"))

}

df_user_info$yelping_since = ymd(df_user_info$yelping_since)



#rm(list = c("c_chunk", "data_chunk", "i", "j"))

print("dataframe for basic user information created")

save(df_user_info, file = "dataset/user_info.Rdata")

print("dataframe for basic user information saved to an Rdata file")





################################################

# create a dataframe regarding friends information

df_user_friends = data.frame(df_user_info$user_id, df_user_info$n_friends, friends_id,

                             stringsAsFactors = F)

colnames(df_user_friends) = c("user_id", "n_friends", "friends_id")

print("dataframe for friends information created")

save(df_user_friends, file = "dataset/user_friends.Rdata")

print("dataframe for friends information saved to an Rdata file")





# ################################################
# 
# # create a dataframe regarding elite status
# 
# year_earlist = min(min_elite, na.rm = T)
# 
# year_latest = max(max_elite, na.rm = T)
# 
# cname_elite = paste0("Y", year_earlist:year_latest)
# 
# 
# 
# # create a function to show elite status of each user for each year
# 
# find_elite_status = function(year){
# 
#   es = function(row_user){
# 
#     if(grepl(pattern = year, row_user)){
# 
#       return("yes")
# 
#     } else return("no")
# 
#   }
# 
#   return(sapply(elite_years, es))
# 
# }
# 
# 
# 
# df_user_elite = data.frame(df_user_info$user_id, stringsAsFactors = F)
# 
# for (i in year_earlist:year_latest){
# 
#   elite_status = find_elite_status(i)
# 
#   elite_status = as.factor(elite_status)
# 
#   df_user_elite = data.frame(df_user_elite, elite_status, stringsAsFactors = F)
# 
# }
# 
# colnames(df_user_elite) = c("user_id", cname_elite)
# 
# print("dataframe for elite status created")
# 
# save(df_user_elite, file = "dataset/user_elite.Rdata")
# 
# print("dataframe for elite information saved to an Rdata file")
```
